move generic stream headers
